Split ev_op init and changed imports
