+++ src/eko/evolution_operator/__init__.py
@@ -11,599 +11,20 @@ from dataclasses import dataclass
 from multiprocessing import Pool
 from typing import Dict, Tuple
 
-import numba as nb
 import numpy as np
 from scipy import integrate
 
-import ekore.anomalous_dimensions.polarized.space_like as ad_ps
-import ekore.anomalous_dimensions.unpolarized.space_like as ad_us
-import ekore.anomalous_dimensions.unpolarized.time_like as ad_ut
-
+from .quad_ker import quad_ker
 from .. import basis_rotation as br
-from .. import interpolation, mellin
 from .. import scale_variations as sv
 from ..couplings import Couplings
 from ..interpolation import InterpolatorDispatcher
 from ..io.types import EvolutionMethod, OperatorLabel
 from ..kernels import ev_method
-from ..kernels import non_singlet as ns
-from ..kernels import non_singlet_qed as qed_ns
-from ..kernels import singlet as s
-from ..kernels import singlet_qed as qed_s
-from ..kernels import valence_qed as qed_v
-from ..matchings import Atlas, Segment, lepton_number
+from ..matchings import Atlas, Segment
 from ..member import OpMember
-from ..scale_variations import expanded as sv_expanded
-from ..scale_variations import exponentiated as sv_exponentiated
 
 logger = logging.getLogger(__name__)
-
-
-@nb.njit(cache=True)
-def select_singlet_element(ker, mode0, mode1):
-    """Select element of the singlet matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        singlet integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-
-    Returns
-    -------
-    complex
-        singlet integration kernel element
-    """
-    j = 0 if mode0 == 100 else 1
-    k = 0 if mode1 == 100 else 1
-    return ker[j, k]
-
-
-@nb.njit(cache=True)
-def select_QEDsinglet_element(ker, mode0, mode1):
-    """Select element of the QEDsinglet matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        QEDsinglet integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-    Returns
-    -------
-    ker : complex
-        QEDsinglet integration kernel element
-    """
-    if mode0 == 21:
-        index1 = 0
-    elif mode0 == 22:
-        index1 = 1
-    elif mode0 == 100:
-        index1 = 2
-    else:
-        index1 = 3
-    if mode1 == 21:
-        index2 = 0
-    elif mode1 == 22:
-        index2 = 1
-    elif mode1 == 100:
-        index2 = 2
-    else:
-        index2 = 3
-    return ker[index1, index2]
-
-
-@nb.njit(cache=True)
-def select_QEDvalence_element(ker, mode0, mode1):
-    """Select element of the QEDvalence matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        QEDvalence integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-    Returns
-    -------
-    ker : complex
-        QEDvalence integration kernel element
-    """
-    index1 = 0 if mode0 == 10200 else 1
-    index2 = 0 if mode1 == 10200 else 1
-    return ker[index1, index2]
-
-
-spec = [
-    ("is_singlet", nb.boolean),
-    ("is_QEDsinglet", nb.boolean),
-    ("is_QEDvalence", nb.boolean),
-    ("is_log", nb.boolean),
-    ("logx", nb.float64),
-    ("u", nb.float64),
-]
-
-
-@nb.experimental.jitclass(spec)
-class QuadKerBase:
-    """Manage the common part of Mellin inversion integral.
-
-    Parameters
-    ----------
-    u : float
-        quad argument
-    is_log : boolean
-        is a logarithmic interpolation
-    logx : float
-        Mellin inversion point
-    mode0 : str
-        first sector element
-    """
-
-    def __init__(self, u, is_log, logx, mode0):
-        self.is_singlet = mode0 in [100, 21, 90]
-        self.is_QEDsinglet = mode0 in [21, 22, 100, 101, 90]
-        self.is_QEDvalence = mode0 in [10200, 10204]
-        self.is_log = is_log
-        self.u = u
-        self.logx = logx
-
-    @property
-    def path(self):
-        """Return the associated instance of :class:`eko.mellin.Path`."""
-        if self.is_singlet or self.is_QEDsinglet:
-            return mellin.Path(self.u, self.logx, True)
-        else:
-            return mellin.Path(self.u, self.logx, False)
-
-    @property
-    def n(self):
-        """Returs the Mellin moment :math:`N`."""
-        return self.path.n
-
-    def integrand(
-        self,
-        areas,
-    ):
-        """Get transformation to Mellin space integral.
-
-        Parameters
-        ----------
-        areas : tuple
-            basis function configuration
-
-        Returns
-        -------
-        complex
-            common mellin inversion integrand
-        """
-        if self.logx == 0.0:
-            return 0.0
-        pj = interpolation.evaluate_grid(self.path.n, self.is_log, self.logx, areas)
-        if pj == 0.0:
-            return 0.0
-        return self.path.prefactor * pj * self.path.jac
-
-
-@nb.njit(cache=True)
-def quad_ker(
-    u,
-    order,
-    mode0,
-    mode1,
-    method,
-    is_log,
-    logx,
-    areas,
-    as_list,
-    mu2_from,
-    mu2_to,
-    a_half,
-    alphaem_running,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-    n3lo_ad_variation,
-    is_polarized,
-    is_time_like,
-    use_fhmruvv,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    u : float
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    is_log : boolean
-        is a logarithmic interpolation
-    logx : float
-        Mellin inversion point
-    areas : tuple
-        basis function configuration
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    mu2_from : float
-        initial value of mu2
-    mu2_from : float
-        final value of mu2
-    aem_list : list
-        list of electromagnetic coupling values
-    alphaem_running : bool
-        whether alphaem is running or not
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an intermediate threshold operator?
-    n3lo_ad_variation : tuple
-        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
-    is_polarized : boolean
-        is polarized evolution ?
-    is_time_like : boolean
-        is time-like evolution ?
-    use_fhmruvv : bool
-        if True use the |FHMRUVV| |N3LO| anomalous dimension
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    ker_base = QuadKerBase(u, is_log, logx, mode0)
-    integrand = ker_base.integrand(areas)
-    if integrand == 0.0:
-        return 0.0
-    if order[1] == 0:
-        ker = quad_ker_qcd(
-            ker_base,
-            order,
-            mode0,
-            mode1,
-            method,
-            as_list[-1],
-            as_list[0],
-            nf,
-            L,
-            ev_op_iterations,
-            ev_op_max_order,
-            sv_mode,
-            is_threshold,
-            is_polarized,
-            is_time_like,
-            n3lo_ad_variation,
-            use_fhmruvv,
-        )
-    else:
-        ker = quad_ker_qed(
-            ker_base,
-            order,
-            mode0,
-            mode1,
-            method,
-            as_list,
-            mu2_from,
-            mu2_to,
-            a_half,
-            alphaem_running,
-            nf,
-            L,
-            ev_op_iterations,
-            ev_op_max_order,
-            sv_mode,
-            is_threshold,
-            n3lo_ad_variation,
-            use_fhmruvv,
-        )
-
-    # recombine everything
-    return np.real(ker * integrand)
-
-
-@nb.njit(cache=True)
-def quad_ker_qcd(
-    ker_base,
-    order,
-    mode0,
-    mode1,
-    method,
-    as1,
-    as0,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-    is_polarized,
-    is_time_like,
-    n3lo_ad_variation,
-    use_fhmruvv,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    quad_ker : float
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an itermediate threshold operator?
-    n3lo_ad_variation : tuple
-        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
-    use_fhmruvv : bool
-        if True use the |FHMRUVV| |N3LO| anomalous dimensions
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    # compute the actual evolution kernel for pure QCD
-    if ker_base.is_singlet:
-        if is_polarized:
-            if is_time_like:
-                raise NotImplementedError("Polarized, time-like is not implemented")
-            else:
-                gamma_singlet = ad_ps.gamma_singlet(order, ker_base.n, nf)
-        else:
-            if is_time_like:
-                gamma_singlet = ad_ut.gamma_singlet(order, ker_base.n, nf)
-            else:
-                gamma_singlet = ad_us.gamma_singlet(
-                    order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
-                )
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_singlet = sv_exponentiated.gamma_variation(
-                gamma_singlet, order, nf, L
-            )
-        ker = s.dispatcher(
-            order,
-            method,
-            gamma_singlet,
-            as1,
-            as0,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv_expanded.singlet_variation(gamma_singlet, as1, order, nf, L, dim=2)
-            ) @ np.ascontiguousarray(ker)
-        ker = select_singlet_element(ker, mode0, mode1)
-    else:
-        if is_polarized:
-            if is_time_like:
-                raise NotImplementedError("Polarized, time-like is not implemented")
-            else:
-                gamma_ns = ad_ps.gamma_ns(order, mode0, ker_base.n, nf)
-        else:
-            if is_time_like:
-                gamma_ns = ad_ut.gamma_ns(order, mode0, ker_base.n, nf)
-            else:
-                gamma_ns = ad_us.gamma_ns(
-                    order, mode0, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
-                )
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_ns = sv_exponentiated.gamma_variation(gamma_ns, order, nf, L)
-        ker = ns.dispatcher(
-            order,
-            method,
-            gamma_ns,
-            as1,
-            as0,
-            nf,
-        )
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = sv_expanded.non_singlet_variation(gamma_ns, as1, order, nf, L) * ker
-    return ker
-
-
-@nb.njit(cache=True)
-def quad_ker_qed(
-    ker_base,
-    order,
-    mode0,
-    mode1,
-    method,
-    as_list,
-    mu2_from,
-    mu2_to,
-    a_half,
-    alphaem_running,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-    n3lo_ad_variation,
-    use_fhmruvv,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    ker_base : QuadKerBase
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    mu2_from : float
-        initial value of mu2
-    mu2_from : float
-        final value of mu2
-    aem_list : list
-        list of electromagnetic coupling values
-    alphaem_running : bool
-        whether alphaem is running or not
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an itermediate threshold operator?
-    n3lo_ad_variation : tuple
-        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
-    use_fhmruvv : bool
-        if True use the |FHMRUVV| |N3LO| anomalous dimensions
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    # compute the actual evolution kernel for QEDxQCD
-    if ker_base.is_QEDsinglet:
-        gamma_s = ad_us.gamma_singlet_qed(
-            order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
-        )
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_s = sv_exponentiated.gamma_variation_qed(
-                gamma_s, order, nf, lepton_number(mu2_to), L, alphaem_running
-            )
-        ker = qed_s.dispatcher(
-            order,
-            method,
-            gamma_s,
-            as_list,
-            a_half,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        # TODO : in this way a_half[-1][1] is the aem value computed in
-        # the middle point of the last step. Instead we want aem computed in mu2_final.
-        # However the distance between the two is very small and affects only the running aem
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv_expanded.singlet_variation_qed(
-                    gamma_s, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-            ) @ np.ascontiguousarray(ker)
-        ker = select_QEDsinglet_element(ker, mode0, mode1)
-    elif ker_base.is_QEDvalence:
-        gamma_v = ad_us.gamma_valence_qed(
-            order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
-        )
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_v = sv_exponentiated.gamma_variation_qed(
-                gamma_v, order, nf, lepton_number(mu2_to), L, alphaem_running
-            )
-        ker = qed_v.dispatcher(
-            order,
-            method,
-            gamma_v,
-            as_list,
-            a_half,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv_expanded.valence_variation_qed(
-                    gamma_v, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-            ) @ np.ascontiguousarray(ker)
-        ker = select_QEDvalence_element(ker, mode0, mode1)
-    else:
-        gamma_ns = ad_us.gamma_ns_qed(
-            order, mode0, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
-        )
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_ns = sv_exponentiated.gamma_variation_qed(
-                gamma_ns, order, nf, lepton_number(mu2_to), L, alphaem_running
-            )
-        ker = qed_ns.dispatcher(
-            order,
-            method,
-            gamma_ns,
-            as_list,
-            a_half[:, 1],
-            alphaem_running,
-            nf,
-            ev_op_iterations,
-            mu2_from,
-            mu2_to,
-        )
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = (
-                sv_expanded.non_singlet_variation_qed(
-                    gamma_ns, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-                * ker
-            )
-    return ker
 
 
 OpMembers = Dict[OperatorLabel, OpMember]
